@@ -1,12 +1,12 @@
 package com.br.swile.tech.model
 
-import java.util.Date
+import kotlinx.datetime.Instant
 
 data class Transaction(
     val id: String,
     val description: String,
     val type: TransactionType,
-    val date: Date,
+    val date: Instant,
     val amount: Double,
     val currency: Currency,
     val smallIcon: Icon,
